Allow to define %SGE job ...% and %SLURM job ...% templates.
Both templates can be combined in a single file and selected by the
%jobmanager ...% option within the file, or by command line using the
--jobmanager (-j) option. (If it exists, a command line argument
overrides the setting in the file.)

Using the %exec ...% block and %eval ...% in the job templates, common
variables can be defined and used.

Backward compatibility is maintained and the generic %job ...% template
is still supported and used if no specific %[SGE/SLURM] job ...% template
is specified.