Fixed an issue in the particle creator classes regarding the computation of particlesPerCell.
The lattice spacing "as" is always supposed to be an even multiple of the time-step "at" in order for the particle code to work.
The number of particles per cell is then computed as (int) "as/at". This division of doubles and subsequent casting to int can sometimes give "wrong" results due to floating-point arithmetic, such as: (int) 0.594603557501/0.297301778751 = 1 instead of 2.
Replacing the cast with Math.round fixes this problem.